Extract NodeLogDirectoryProvider to electron specific file
