+++ extensions/typescript-language-features/src/utils/logDirectoryProvider.ts
@@ -3,38 +3,12 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as fs from 'fs';
-import * as path from 'path';
-import * as vscode from 'vscode';
-import { memoize } from './memoize';
+export interface ILogDirectoryProvider {
+	getNewLogDirectory(): string | undefined;
+}
 
-export default class LogDirectoryProvider {
-	public constructor(
-		private readonly context: vscode.ExtensionContext
-	) { }
-
-	public getNewLogDirectory(): string | undefined {
-		const root = this.logDirectory();
-		if (root) {
-			try {
-				return fs.mkdtempSync(path.join(root, `tsserver-log-`));
-			} catch (e) {
-				return undefined;
-			}
-		}
+export const noopLogDirectoryProvider = new class implements ILogDirectoryProvider {
+	public getNewLogDirectory(): undefined {
 		return undefined;
 	}
-
-	@memoize
-	private logDirectory(): string | undefined {
-		try {
-			const path = this.context.logPath;
-			if (!fs.existsSync(path)) {
-				fs.mkdirSync(path);
-			}
-			return this.context.logPath;
-		} catch {
-			return undefined;
-		}
-	}
-}
+};
